Fix cyclic import
